requirements.txt
flask
requests
Pillow  # For image processing
azure-cli
python-dotenv
black